pyproject.toml
[tool.poetry]
name = "bookstore-work"
version = "0.1.0"
description = "Bookstorework API"
authors = ["Samuel Ivanaskas <samuelivanaskas@hotmail.com>"]
readme = "README.md"
packages = [{include = "bookstore_work"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.0"
factory-boy = "^3.3.0"
django = "^4.2.3"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
